+++ R/ROADMAP_construct_reference.R
@@ -4,24 +4,26 @@
 #' @param keyword_query Search all columns in the Roadmap annotations metadata
 #' and only query annotations that contain your keywords.
 #' Can provide multiple keywords in list form:
-#' \code{c("placenta","liver","monocytes")}
-#'
+#' \code{c("placenta","liver","monocytes")}.
+#' @param verbose Print messages.
+#' @family ROADMAP
+#' 
+#' @export
+#' @importFrom data.table transpose fread
 #' @examples
 #' ref <- ROADMAP_construct_reference(keyword_query = c(
 #'     "placenta",
 #'     "liver",
 #'     "monocytes"
 #' ))
-#' @family ROADMAP
-#' @export
-#' @importFrom data.table transpose fread
 ROADMAP_construct_reference <- function(ref_path =
                                             system.file(
                                                 "extdata/ROADMAP",
                                                 "ROADMAP_Epigenomic.js",
                                                 package = "echoannot"
                                             ),
-                                        keyword_query = NULL) {
+                                        keyword_query = NULL,
+                                        verbose = TRUE) {
     # %like% is from data.table
     ref <- suppressWarnings(data.table::fread(ref_path))
     colnames(ref)[1] <- "EID"
@@ -32,8 +34,10 @@ ROADMAP_construct_reference <- function(ref_path =
         )
         ref <- ref[rows, ]
         messager(
-            "+ ROADMAP::", nrow(ref),
-            "annotation(s) identified that match `keyword_query`."
+            "+ ROADMAP::", formatC(nrow(ref),big.mark = ","),
+            "annotation(s) identified that match:",
+            paste(keyword_query,collapse = " | "),
+            v=verbose
         )
     }
     return(ref)