+++ R/XGR_import_annotations.R
@@ -32,24 +32,20 @@ XGR_import_annotations <- function(gr.snp,
     }
     GR.orig <- unlist(GR.annotations)
 
-    gr.xgr <- lapply(names(GR.orig), function(g, gr.snp. = gr.snp) {
-        # messager("Finding overlap for:", g)
+    gr.xgr <- lapply(names(GR.orig), function(g, gr.snp. = gr.snp) { 
         subsetByOverlaps <- get(
             "subsetByOverlaps",
             asNamespace("GenomicRanges")
         )
         GR.overlap <- subsetByOverlaps(GR.orig[[g]], gr.snp.)
-        len <- length(seqnames(GR.overlap))
-        # messager("   - Overlapping annotations = ",len)
+        len <- length(seqnames(GR.overlap)) 
         if (len > 0) {
             return(GR.overlap)
         } else {
             return(NULL)
         }
     })
-    grl.xgr <- name_filter_convert(gr.xgr,
-        names(GR.orig),
-        min_hits = annot_overlap_threshold
-    )
+    grl.xgr <- name_filter_convert(grl = gr.xgr,
+                                   min_hits = annot_overlap_threshold)
     return(grl.xgr)
 }